hardware_imu layer changed
